keeper post mapping

+++ keeper-notes/src/main/java/keeper/notes/entity/Animal.java
@@ -20,7 +20,7 @@ public class Animal {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int animalId;
+	private Long animalId;
 	
 	private String species;
 	private String commonName;
